Move components out of housing folders
